Create amcgover_changes.py
The original didn't work on my Windows machine. It was giving the error described here: http://stackoverflow.com/questions/25887081/python-java-pipe-dev-stdin-file-not-found-exception 

I made the following changes and got it working:
_call_runtagger
Run command changed to take a file name instead of list of tweets.
The output of the call to the tagger is read similarly to the way it is read in _check_script_is_present. The 'result' list contains the tagged tweets. Each element in 'result' is a single tagged tweet.
The line delimiting each parsed tweet is '\r\n', not '\n\n'. (On my machine anyway)

check_script_is_present has 'while po.poll()' removed from it